gnu: wxwidgets: Add wxWidgets without EGL.
Detected when working on Kicad-9.  wxwidgets was built with EGL but
Kicad was unable to create the OpenGL context.  According to Debian:

<https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1024147;msg=5>

This happens due to Glew having EGL disabled, and would break packages
that use both `wxwidgets` and `glew` together.  We already detected the
issue with PrusaSlicer, that's why we had a separate package for it, but
it also needed extra unrelated changes.

* gnu/packages/wxwidgets.scm (wxwidgets-sans-egl): Add it.
(prusa-wxwidgets): Inherit from it.

Change-Id: Idde2985501b4d06cbb5bbd92030d5516bf7bea07