gnu: librewolf: Fix mozconfig creation.
When 'configure creates the mozconfig used to build the browser, it overwrites
upstream’s configuration, requiring that it repeat the options already set
upstream.  It also handles the final two options incorrectly, due to missing
newlines -- the options are concatenated.

Instead of doing that all that, append the Guix-specific options to
upstream’s mozconfig.

* gnu/packages/librewolf.scm (librewolf): Fix mozconfig creation.
[#:configure-flags]: Remove flags already present in upstream mozconfig.
[#:configure-flags]: Disable toolchain bootstrapping.
[phase 'configure]: Remove flags already present in upstream mozconfig.
[phase 'configure]: Append to mozconfig instead of overwriting.

Change-Id: I67070ac1e84747ea4f88c527441ffcea4c2e02f5