gnu: fwupd: Add elogind input
The fwupd package doesn't contain the dbus service in
share/dbus-1/system-service, that is because it is installed only
conditionally when elogind is available.

This will introduce the service in the output, so user's won't have to start
fwupd themselves, it will get started by dbus on demand.

I don't think this should break anything for systems that do not use elogin,
the only places this is used will fall back to error:
- https://github.com/fwupd/fwupd/blob/1.8.14/src/fu-util-common.c#L242,
- https://github.com/fwupd/fwupd/blob/1.8.14/src/fu-util-common.c#L287
Meaning users without elogind will get error either way, just a different kind
of error.

* gnu/packages/firmware.scm: (fwupd)[inputs]: Add elogind

Change-Id: Id44001de6950bb6b39f1b3bba316a08ce6640fc0
Signed-off-by: Ludovic Courtès <ludo@gnu.org>