gnu: insight-toolkit: Exclude gtest target.
Fixes: guix/guix#776

This new phase makes it possible to configure projects that use
insight-toolkit.  I think this is what the configure flag '-DGTEST_ROOT=gtest'
was intended to do, but it does not.  In addition, the configure flag
'-DITK_USE_SYSTEM_GOOGLETEST=ON' and native input googletest are no longer
required to placate the configure phase.

* gnu/packages/image-processing (insight-toolkit) [arguments]
<#:configure-flags>: Remove -DITK_USE_SYSTEM_GOOGLETEST=ON and
-DGTEST_ROOT=gtest.
<#:phases>: Add phase 'exclude-gtest-target.
[native-inputs]: Remove googletest.

Change-Id: I9c2e051ac9cfc3439ea53e1a4f3207fbb2f3337b
Signed-off-by: Ludovic Courtès <ludo@gnu.org>