tests: Really test the etc-bashrc-d-service-type service.
Follow-up of: 4c017ccfe56a8887bf26faaea62d60501d7cc4f6.

The --init-file script will only be run in the context of an interactive
shell, which is not the case here. Use the `-i` argument instead to force an
interactive shell. That way, the test_bashrc_d.sh is now really executed.

* gnu/tests/base.scm (run-basic-test): Fix the etc-bashrc-d-service-type test.

Change-Id: I3d749f65f51ff103c76f4e3d60746ae963660329