gnu: python-matplotlib: Patch dlopen so MPLBACKEND is not overridden.
Fixes <https://issues.guix.gnu.org/70687>.

In Python, importing 'matplotlib.pyplot' sets the backend used by Matplotlib
to Agg unless it was set to WebAgg or nbAgg.  This is because it incorrectly
determines that no event loop can be started due to a failed dlopen call for
'libX11.so.6'.  This is fixed by patching the dlopen path.

* gnu/packages/python-xyz.scm (python-matplotlib) [arguments]
<#:phases>: Add 'patch-dlopen.

Change-Id: I3e45c9a2921374e3b16cfcbd8e3e1e91780da306
Signed-off-by: Sharlatan Hellseher <sharlatanus@gmail.com>