profiles: Preserve Bash ‘HIST’ variables.
This reduces the risk that, when using Bash, ‘guix shell --pure’ would
spawn an instance of Bash that would ignore the user’s history-related
settings.

It is not bulletproof though: these variables as shell variables that
need not be exported as environment variable.  When they are not
exported, the user’s history-related settings would still be ignored.

* guix/profiles.scm (%precious-variables): Add HIST variables from Bash.

Change-Id: Id2898d67d2e70c083abbeacad3f0663940fb6789