gnu: darktable: Fix build.
Darktable fails to build with "sorry, unimplemented: Graphite loop
optimizations cannot be used (isl is not available)" when not specifying
gcc-14 explicitly in native-inputs and just using the implicit default gcc-14.
It is unclear why.  (With older gcc the build fails due to an ABI
incompatibility as openexr is now built with gcc-14.)

* gnu/packages/photo.scm (darktable)[native-inputs]: Change gcc-13 to gcc-14.

Change-Id: I97dd1a2aaeaaf065dc2a22d20487405c8223cdc0