gnu: Add python-docker-5.
This is a followup to commit 3eb1ceac58985c0407f3ade1cfa72dbee1b7886a along
with an upcoming commit to use this in docker-compose to fully fix it.  This
package is added just for the old version 1 of docker-compose and should be
removed once Docker is updated (which includes "docker compose" v2 built-in).

* gnu/packages/docker.scm (python-docker-5): New variable.

Change-Id: I9e640ca3a87fb6e055524cd5acc79588ac36fe36