installer: wait-service-online: Allow user to proceed if offline.
The user may choose to proceed with installation
even if they seem to be offline. This is mostly for cases
where the user if actually online, but the gnu and guix
infrastructure is down. On top of that, if the installer
has all the necessary dependencies the user might not need
internet at all.

* gnu/installer/newt/network.scm (wait-oservice-online): Add choice window for
user to continue if offline.

Change-Id: Id8868ad42a6874d31b64765457f8a3d3b08ebd6c
Signed-off-by: Ludovic Courtès <ludo@gnu.org>