vm-image.tmpl: Use /dev/vda2 as root file system device.
The /dev/vda1 partition points to an EFI partition.
The /run/current-system/configuration.scm will then point to an incorrect
partition - to the EFI partition instead of root partition. Fix that by
pointing it to the default root-label "Guix_image".
This allows users to reconfigure out of the /run/current-system/configuration.scm.

* gnu/system/examples/vm-image.tmpl: Switch root filesystem device
to /dev/vda2.

Change-Id: I983892fb3c860dc5bbb2a1654d28db83e88d8c1a
Signed-off-by: Ludovic Courtès <ludo@gnu.org>