gnu: python-mohawk: Fix build (for now).
python-mohawk is a leaf package whose source hasn't been updated in six
years.  All of the forks still import six. It seems ready for
deprecation. Nevertheless, as long as we still have six in the
repository, we might as well make the package build. Added a comment
that can be seen when grepping for 'python-six'.

* gnu/packages/python-web.scm (python-mohawk): Fix build.
[propagated-inputs]: Add python-six.

Change-Id: Iaac6c2562e65843946361f1cce94ba38207a9640
Signed-off-by: Sharlatan Hellseher <sharlatanus@gmail.com>